pyproject.toml
[tool.poetry]
name = "testsmth"
version = "0.1.0"
description = ""
authors = ["yarolys <lysov.yaroslav.r@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["all"], version = "^0.115.8"}
sqlalchemy = "^2.0.38"
alembic = "^1.14.1"
asyncpg = "^0.30.0"
passlib = "^1.7.4"
python-jose = "^3.3.0"
pydantic-settings = "^2.7.1"
fastapi-cache2 = "^0.2.2"
redis = "^5.2.1"
orjson = "^3.10.15"
celery = "^5.4.0"
flower = "^2.0.1"
pillow = "^11.1.0"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
